add: status as enum
